refactor: lotto.model 패키지로 이동
