+++ src/main/java/lotto/model/WinningNumber.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.model;
 
 import java.util.List;
+import lotto.controller.LottoRules;
 
 public class WinningNumber {
