+++ src/main/java/lotto/model/WinningNumberValidator.java
@@ -1,7 +1,8 @@
-package lotto;
+package lotto.model;
 
 import java.util.Collections;
 import java.util.List;
+import lotto.controller.LottoRules;
 
 public class WinningNumberValidator {
